Do not force record to have and id when adding it to SC.ManyArray
That way we do not have to do anything in application to
add record to SC.ManyArray immediately after creation,
it will be added just after it gets an id.

Eventually the best approach to deal with SC.ManyArray would be
to rewrite it to work on storeKeys instead of storeIds, which will
allow to lazily load everything without hacks like this one.